% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_garch.R
\name{plotgarch}
\alias{plotgarch}
\title{plot Garch}
\usage{
plotgarch(x, T)
}
\arguments{
\item{x}{an xts format serie}

\item{T}{give a the title for the plot with "" bracket}
}
\value{
Return a plot for a Garch model
}
\description{
plot Garch
}
