% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_garch.R
\name{plotgarch}
\alias{plotgarch}
\title{plot garch model}
\usage{
plotgarch(x, T)
}
\arguments{
\item{x}{the data set in xts format}

\item{T}{give a the title for the plot}
}
\value{

}
\description{
plot garch model
}
